同步官方最新代码 | 新增各CPU核自主管理中断     百图画鸿蒙 + 百文说内核 + 百万注源码  => 挖透鸿蒙内核源码     鸿蒙研究站 | http://weharmonyos.com (国内)               | https://weharmony.github.io (国外)     oschina | https://my.oschina.net/weharmony     博客园 | https://www.cnblogs.com/weharmony/     知乎 | https://www.zhihu.com/people/weharmonyos     csdn | https://blog.csdn.net/kuangyufei     51cto | https://harmonyos.51cto.com/column/34     掘金 | https://juejin.cn/user/756888642000808     公众号 | 鸿蒙研究站 (weharmonyos)

+++ arch/arm/arm/src/strncpy_from_user.c
@@ -49,7 +49,7 @@ INT32 LOS_StrncpyFromUser(CHAR *dst, const CHAR *src, INT32 count)
     }
 
     maxCount = (LOS_IsUserAddressRange((VADDR_T)(UINTPTR)src, (size_t)count)) ? \
-                count : (USER_ASPACE_TOP_MAX - (UINTPTR)src);//最大能拷贝的数据量,结束地址不能超过 USER_ASPACE_TOP_MAX
+                count : (INT32)(USER_ASPACE_TOP_MAX - (UINTPTR)src);//最大能拷贝的数据量,结束地址不能超过 USER_ASPACE_TOP_MAX
 															//USER_ASPACE_TOP_MAX 是用户空间能触及的最大虚拟内存空间地址
     for (i = 0; i < maxCount; ++i) {//一个个字符拷贝
         if (LOS_GetUser(&character, src + offset) != LOS_OK) {
